Allow more than one query element if extended CSS selector is used.
If extended CSS selectors are used in HTMX attributes (e.g. hx-include ), only the first extended CSS selector is processed.
And further the extended selector is processed only if it is the first selector.
This commit splits the attribute at commas and then process all the CSS selectors one by one.